requirements.txt
opencv-python>=4.8.0
Pillow>=9.5.0
ffmpeg-python>=0.2.0
python-dotenv>=1.0.0
tqdm>=4.65.0
colorama>=0.4.6 